backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
playwright==1.41.2
beautifulsoup4==4.12.3
weaviate-client==4.4.2
redis==5.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
sentence-transformers==2.3.1
torch==2.1.2
